fix(runtime-html): keep application lifecycle failures terminal
