+++ packages/runtime-html/src/errors.ts
@@ -168,13 +168,7 @@ export const enum ErrorNames {
   hydration_target_count_mismatch = 822,
   hydration_node_count_mismatch = 823,
   hydration_view_count_mismatch = 824,
-  app_root_deactivation_cleanup_failed = 825,
   app_task_phase_failed = 826,
-  aurelia_standalone_root_rollback_failed = 827,
-  aurelia_queued_start_disposal_failed = 828,
-  aurelia_start_rollback_failed = 829,
-  aurelia_stop_vetoed = 830,
-  aurelia_stop_cleanup_failed = 831,
   au_compose_operation_teardown_failed = 832,
   au_compose_deactivation_disposal_failed = 833,
   au_compose_activation_disposal_failed = 834,
@@ -330,13 +324,7 @@ const errorsMap: Record<ErrorNames, string> = {
   [ErrorNames.hydration_target_count_mismatch]: `SSR hydration error: manifest declares {{0}} targets but collected {{1}} from DOM.`,
   [ErrorNames.hydration_node_count_mismatch]: `SSR hydration error: manifest declares {{0}} total nodes for views but found {{1}} nodes in DOM.`,
   [ErrorNames.hydration_view_count_mismatch]: `SSR hydration error: manifest declares {{0}} views but items array has {{1}} elements.`,
-  [ErrorNames.app_root_deactivation_cleanup_failed]: `App root deactivation failed during cleanup`,
-  [ErrorNames.app_task_phase_failed]: `Application task phase failed`,
-  [ErrorNames.aurelia_standalone_root_rollback_failed]: `Standalone Aurelia root activation failed during rollback`,
-  [ErrorNames.aurelia_queued_start_disposal_failed]: `Queued Aurelia start failed during disposal`,
-  [ErrorNames.aurelia_start_rollback_failed]: `Aurelia start failed during rollback`,
-  [ErrorNames.aurelia_stop_vetoed]: `Aurelia stop was vetoed during application deactivation`,
-  [ErrorNames.aurelia_stop_cleanup_failed]: `Aurelia stop failed during cleanup`,
+  [ErrorNames.app_task_phase_failed]: `Multiple application tasks failed during the "{{0}}" phase.`,
   [ErrorNames.au_compose_operation_teardown_failed]: `AuCompose operation failed during teardown cleanup`,
   [ErrorNames.au_compose_deactivation_disposal_failed]: `AuCompose deactivation failed during disposal`,
   [ErrorNames.au_compose_activation_disposal_failed]: `AuCompose activation failed during disposal`,
